requirements.txt
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
jinja2>=3.1.0
pyyaml>=6.0.1
